Removed TStr and replaced with global operator implicit string concatenation.
Added XML files.

Changed Engine.Dat to Engine.XML.

Working to change .dat files to more readable XML format.

Added base layers for Vulkan rendering.

+++ src/base/FileSystem.h
@@ -72,7 +72,7 @@ public:
 
 
     static void SDLFileFree(char*& pOutData);
-    static int SDLFileRead(std::string fname, char*& pOutData, uint32_t& _iOutSizeBytes, bool addNull = false);
+    static int SDLFileRead(std::string fname, char*& pOutData, int64_t& _iOutSizeBytes, bool addNull = false);
     static int SDLFileWrite(std::string fname, char* pData, uint32_t _iDataSizeBytes);
     static std::string getFilePath(std::string name);
